requirements.txt
Flask==3.0.3
PyJWT==2.9.0
RPi.GPIO==0.7.1
